+++ CHANGELOG.md
@@ -4,7 +4,12 @@
 
 ### Added
 
+- Created the `ModuleConfig` class that wraps command config data for future configuration expansions.
 - The initial `run.py` file now checks for any import errors. If detected, it'll try to install the requirements file.
+
+### Changed
+
+- All instances of `plugin` have been renamed to `module`. Meaning that you need to rename `config/plugins` to `config/modules`.
 
 ## 2019-02-12